Add BERT example w/o data loader

+++ texar/models/transformer.py
@@ -138,7 +138,7 @@ class Transformer(ModuleBase):
             return mle_loss
         else:
             start_tokens = torch.full([batch_size], self.bos_token_id,
-                                      dtype=torch.int64)
+                                      dtype=torch.int32)
 
             if torch.cuda.is_available():
                 start_tokens = start_tokens.cuda()
@@ -217,6 +217,7 @@ class LabelSmoothingLoss(nn.Module):
             output.reshape(ori_shapes[0]),
             model_prob.reshape(ori_shapes[0]),
         )
+
         return sequence_softmax_cross_entropy(
             labels=model_prob, logits=output, sequence_length=label_lengths,
             average_across_batch=False, sum_over_timesteps=False,
